fix for retrieving env variable in client component when it should be server component
